feat: change ErrorPage render logic
