+++ src/views/ErrorPage/ErrorPage.module.scss
@@ -19,6 +19,7 @@
 .title {
   font-style: $primary-font;
   font-size: 70px;
+  text-align: center;
   color: $primary-text-color;
   text-shadow: 0px 20px 30px rgba(9, 30, 63, 0.1);
