--- src/views/Page404/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Page404';
